Refactor code to use Vue SFC instead of TS files.
