+++ demo-vue/app/examples/Ripples.vue
@@ -1,0 +1,35 @@
+<template>
+  <Page>
+    <ActionBar :title="title">
+      <NavigationButton text="Back"
+                        android.systemIcon="ic_menu_back"
+                        @tap="onNavigationButtonTap"></NavigationButton>
+    </ActionBar>
+    <StackLayout>
+      <MDRipple backgroundColor="gray" rippleColor="green" width="100" height="100"/>
+    </StackLayout>
+  </Page>
+</template>
+
+<script lang="ts">
+import * as frameModule from 'tns-core-modules/ui/frame';
+import Vue from 'vue'
+
+export const title = 'Ripples sample';
+
+export default Vue.extend({
+    name: 'Ripples',
+
+    data() {
+        return {
+            title: title
+        };
+    },
+
+    methods: {
+        onNavigationButtonTap() {
+            frameModule.topmost().goBack();
+        }
+    },
+})
+</script>
