+++ demo-vue/app/examples/Sliders.vue
@@ -1,12 +1,5 @@
-import * as frameModule from 'tns-core-modules/ui/frame';
-
-const description = 'Sliders sample';
-
-export default {
-    name: 'Sliders',
-    description: description,
-    template: `
-    <Page>
+<template>
+  <Page>
       <ActionBar :title="title">
         <NavigationButton text="Back"
                           android.systemIcon="ic_menu_back"
@@ -16,7 +9,7 @@ export default {
         <Label text="Default slider with dynamic value" padding="10" />
         <MDSlider elevation="3" rippleColor="blue"
                   minValue="0" maxValue="100"
-                  v-model="value" @valueChange="onValueChanged" />
+                  v-model="value" />
         <Label text="Green slider without hollow at 0" padding="10" />
         <MDSlider thumbHollowAtStart="false" color="green"
                   minValue="0" maxValue="100"
@@ -26,20 +19,32 @@ export default {
                   minValue="0" maxValue="100"
                   v-model="value" />
       </StackLayout>
-    </Page>
-    `,
-    data() {
+  </Page>
+</template>
+
+<script lang="ts">
+import * as frameModule from 'tns-core-modules/ui/frame';
+import Vue from 'vue';
+
+export const title = 'Sliders sample';
+
+export default Vue.extend({
+    name: 'Sliders',
+    data () {
         return {
-            title: description,
-            value: 50
-        };
+            name: 'Sliders',
+            title: title,
+            value: 50,
+        }
     },
     methods: {
         onNavigationButtonTap() {
             frameModule.topmost().goBack();
         },
+
         onValueChanged({ value }) {
             console.log(`Value changed to ${value}`);
-        }
-    }
-};
+        },
+    },
+})
+</script>
