+++ demo-vue/app/examples/index.ts
@@ -1,11 +1,39 @@
-import ActivityIndicators from './ActivityIndicators';
-import Buttons from './Buttons';
-import CardViews from './CardViews';
-import ProgressBars from './ProgressBars';
-import Ripples from './Ripples';
-import Sliders from './Sliders';
-import TextFields from './TextFields';
+import {title as activityIndicatorsTitle} from './ActivityIndicators.vue';
+import ActivityIndicators from './ActivityIndicators.vue';
+import {title as buttonsTitle} from './Buttons.vue';
+import Buttons from './Buttons.vue';
+import {title as cardViewsTitle} from './CardViews.vue';
+import CardViews from './CardViews.vue';
+import {title as progressBarsTitle} from './ProgressBars.vue';
+import ProgressBars from './ProgressBars.vue';
+import {title as ripplesTitle} from './Ripples.vue';
+import Ripples from './Ripples.vue';
+import {title as slidersTitle} from './Sliders.vue';
+import Sliders from './Sliders.vue';
+import {title as textFieldsTitle} from './TextFields.vue';
+import TextFields from './TextFields.vue';
 
 export const getExamples = () => {
-    return [Buttons, CardViews, ActivityIndicators, TextFields, ProgressBars, Sliders, Ripples];
+    return [{
+        title: buttonsTitle,
+        component: Buttons,
+     }, {
+        title: cardViewsTitle,
+        component: CardViews
+     }, {
+        title: activityIndicatorsTitle,
+        component: ActivityIndicators
+    }, {
+        title: textFieldsTitle,
+        component: TextFields,
+    }, {
+        title: progressBarsTitle,
+        component: ProgressBars,
+    }, {
+        title: slidersTitle,
+        component: Sliders,
+    }, {
+        title: ripplesTitle,
+        component: Ripples,
+    }];
 };
