+++ demo-vue/vue-shims.d.ts
@@ -1,0 +1,5 @@
+declare module "*.vue" {
+  import Vue from "vue";
+  export const title;
+  export default Vue;
+}
