+++ tsconfig.json
@@ -20,7 +20,6 @@
         "noFallthroughCasesInSwitch": true,
         "baseUrl": ".",
         "outDir": "./plugin",
-        "rootDir": "src",
         "paths": {
             "nativescript-material-dialogs": ["src/dialogs/dialogs"],
             "nativescript-material-button": ["src/button/button"],
@@ -32,7 +31,7 @@
             "*": ["node_modules/tns-core-modules/*", "node_modules/*"]
         }
     },
-    "include": ["src/**/*.ts", "./references.d.ts", "./src/references.d.ts"],
+    "include": ["src/**/*", "src/**/*.ts", "./references.d.ts", "./src/references.d.ts"],
     "exclude": ["node_modules", "platforms"],
     "compileOnSave": false
 }
